Git != GitHub
While most of us use GitHub as a primary remote, It shouldn't be taught as
the *reason* git is awesome. There are plenty of workflows that are possible
only outside of the github ecosystem, and the features we're selling here
-- specifically the distributed part of DVCS -- are something that git does
well on its own, so it makes sense to mention them as features and *then*
introduce GitHub as a service that adds value to these features.