Avoid contradiction of terms.
We are hammering into the student that everything is an Object, and then
bam, we hit them in the face with "except these things", which doesn't make
much sense because we are talking two different things.

Everything in the Ruby VM actually is an Object (or BasicObject, if you
want to get technical about it), but we aren't *in* the Ruby VM here, we
are talking *to* it. And yes, we use String literals and Integer literals,
and a lot of the things we are working with have a 1:1 mapping to actual
objects in the Ruby VM, but the conflation of terms becomes a contradiction
with what we just said, and the subtleties of context will likely be lost
on a beginner.